[FIX] log an error on reentrant call to the statecharts
Ask the dev to use the new sc_queue() method.

This has the effect of logging an error on all write() calls in the test
suite. Conclusion is that making write a statechart event is really a
bad idea. So adapt the test suite to implement the guard on write using
a check_write event.