Make sc_*_allowed fields after patching event methods
The initialization of the dummy interpreter that
is necessary to compute the default values for the sc_*_allowed
fields implies entering the initial states. The on entry of
the initial state may execute code that involves calling some
method.origin. This changes makes it so that method.origin
is correctly initialized when the default values are computed.

Fixes #11