mig 11: major simplification
We killed the statechart model, replaced by a class attribute
_statechart_file to declare the yaml to be used; the statechart
model did not work in v11 because in the test suite it was loaded
after loading the view, and the view need fields that are injected
by the statechart; moreover the statechart model gave the impression
one could change it without restarting the server, which was very
far from reality. This makes the code much much simpler.