[FIX] we don't need the separate statchart_injector anymore.
While it was working in -u mode where all python modules were
preloaded before calling _prepare_setup, in normal mode, it could
be running while not all addons were imported.

Moreover the reason I did it that way initially does not exists anymore,
since the _statechart is now stored in the original model class,
and _prepare_setup/_setup_complete do nothing if it is present, so it's
garanteed to run only once.

Moreover, the model class holding _statechart_file is not guaranteed to
be type(seld).__bases__[0], so we look at all base classes to find it.