[IMP] restore possibility to override statechart
We add sc_event_allowed fields of all declared statecharts,
avoiding create the same field twice. This is fine, because
at runtime they will evaluate against the current statechart
(that may have been replaced by a subclass) and the worse that
can happen is that they evaluate to false if the replacing state
chart does not have the corresponding events.

We track if methods have been added/patched using a set (instead
of a per class flag before), so a method will never be patched
more than once. If an event has been removed in an overriding statechart
this is fine: the method will simply raise NoTransitionError at runtime.